Reformat the cleanup task. Fixing a bunch of problematic uses of PendingHandleOp which could have caused memory corruption issues. Now we'll get lint warnings when they are unused, and panics when they are dropped unexpectedly

+++ kernel/src/task/args.rs
@@ -1,5 +1,20 @@
+//! Arguments passed to new executable tasks
+//!
+//! When a task is created, it begin in an empty state with no runnable code.
+//! Only after a program has been loaded and attached is it possible to run
+//! the task. During this intermediate period, it is also possible to pass a
+//! series of arguments to the task.
+//! The `ExecArgs` struct on a Task object uses the builder model. Args can
+//! be added incrementally, and when the task begins to execute the final state
+//! is compiled. These values are placed on the initial stack of the task, where
+//! they can be read by a program, similar to the `argv` array in POSIX/C.
+
 use alloc::vec::Vec;
 
+/// Args are treated externally as a set of distinct strings. Internally we
+/// store them as a continuous array of bytes, with individual lengths marking
+/// the barrier between arguments. This is the format that will be copied to
+/// the task stack when it first executes.
 pub struct ExecArgs {
     raw: Vec<u8>,
     lengths: Vec<u32>,
@@ -13,6 +28,8 @@ impl ExecArgs {
         }
     }
 
+    /// Append a new argument to the list, storing it in a way that will be
+    /// easily copied to the stack later.
     pub fn add(&mut self, arg: &str) -> &mut Self {
         self.lengths.push(arg.len() as u32 + 1);
         for b in arg.bytes() {
@@ -34,11 +51,16 @@ impl ExecArgs {
         self.lengths.len() as u32
     }
 
-    pub fn stack_size(&self) -> u32 {
+    /// Determine how much stack space is required to copy the argument
+    /// information to the task's stack.
+    /// The actual argument bytes will be zero-padded to a 4-byte boundary.
+    /// The lengths will be copied as a series of 32-bit integers, and the total
+    /// count will also be a 32-bit integer.
+    pub fn stack_size(&self) -> usize {
         let mut string_length = self.raw.len();
         if string_length & 3 != 0 {
             string_length += 4 - (string_length & 3);
         }
-        (string_length + self.lengths.len() * 4 + 4) as u32
+        string_length + self.lengths.len() * 4 + 4
     }
 }
